# Conflicts:
#	src/client/BL/Accountant/FinancialAccountbl/FinancialAccountController.java
#	src/client/Vo/coVO.java

+++ src/server/Po/consumerPO.java
@@ -4,15 +4,15 @@ import java.io.Serializable;
 
 public class consumerPO implements Serializable{
 	private String keyno;
-	private Long kinds;
-	private Long lev;
+	private double kinds;
+	private double lev;
 	private String keyname;
 	private String tel;
 	private String mailcode;
 	private String email;
-	private Long capacit;
-	private Long receive;
-	private Long pay;
+	private double capacit;
+	private double receive;
+	private double pay;
 	private String server;
 	private String accoun;
 
@@ -28,19 +28,19 @@ public class consumerPO implements Serializable{
 		this.keyno = keyno;
 	}
 
-	public Long getKinds() {
+	public double getKinds() {
 		return kinds;
 	}
 
-	public void setKinds(Long kinds) {
+	public void setKinds(double kinds) {
 		this.kinds = kinds;
 	}
 
-	public Long getLev() {
+	public double getLev() {
 		return lev;
 	}
 
-	public void setLev(Long lev) {
+	public void setLev(double lev) {
 		this.lev = lev;
 	}
 
@@ -76,27 +76,27 @@ public class consumerPO implements Serializable{
 		this.email = email;
 	}
 
-	public Long getCapacit() {
+	public double getCapacit() {
 		return capacit;
 	}
 
-	public void setCapacit(Long capacit) {
+	public void setCapacit(double capacit) {
 		this.capacit = capacit;
 	}
 
-	public Long getReceive() {
+	public double getReceive() {
 		return receive;
 	}
 
-	public void setReceive(Long receive) {
+	public void setReceive(double receive) {
 		this.receive = receive;
 	}
 
-	public Long getPay() {
+	public double getPay() {
 		return pay;
 	}
 
-	public void setPay(Long pay) {
+	public void setPay(double pay) {
 		this.pay = pay;
 	}
 
@@ -116,7 +116,7 @@ public class consumerPO implements Serializable{
 		this.accoun = accoun;
 	}
 
-	public consumerPO(String keyno, Long kinds, Long lev, String keyname, String tel, String mailcode, String email, Long capacit, Long receive, Long pay, String server, String accoun) {
+	public consumerPO(String keyno, double kinds, double lev, String keyname, String tel, String mailcode, String email, double capacit, double receive, double pay, String server, String accoun) {
 		this.keyno = keyno;
 		this.kinds = kinds;
 		this.lev = lev;
